Kaisong2/issue 59/updateeventsbackend

+++ src/main/java/cs428/project/gather/controllers/EventsController.java
@@ -3,6 +3,8 @@ package cs428.project.gather.controllers;
 import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
+import java.lang.reflect.Type;
+import java.util.Date;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -15,6 +17,11 @@ import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonDeserializationContext;
+import com.google.gson.JsonDeserializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonParseException;
 
 import cs428.project.gather.data.*;
 import cs428.project.gather.data.model.*;
@@ -153,6 +160,127 @@ public class EventsController {
 		return newEvent;
 	}
 	
+	
+	@RequestMapping(value = "/rest/events/update", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
+	@ResponseBody
+	public ResponseEntity<RESTResourceResponseData<Event>> updateEvent(HttpServletRequest request,
+			@RequestBody String rawData, BindingResult bindingResult) {
+		// TODO: Wrap this in TryCatch, report exception to frontend.
+		GsonBuilder builder = new GsonBuilder(); 
+		
+				// Register an adapter to manage the date types as long values 
+		builder.registerTypeAdapter(Date.class, new JsonDeserializer<Date>() { 
+			@Override
+			public Date deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+				return new Date(json.getAsJsonPrimitive().getAsLong()); 
+			}
+		});
+		Gson gson = builder.create();
+		UpdateEventData updateEventData = gson.fromJson(rawData, UpdateEventData.class);
+
+		if (!ActorTypeHelper.isRegisteredUser(request)) {
+			System.out.println("An anonymous user tried to add an event.");
+			bindingResult.reject("-7", "Incorrect User State. Only registered users can add events.");
+			return RESTResourceResponseData.<Event> badResponse(bindingResult);
+		}
+
+		newEventDataValidator.validate(updateEventData, bindingResult);
+		System.out.println("Validated: " + rawData);
+		if (bindingResult.hasErrors()) {
+			return RESTResourceResponseData.<Event> badResponse(bindingResult);
+		}
+
+		Actor actor = ActorStateUtility.retrieveActorFromRequest(request);
+		Registrant owner = this.regRepo.findOne(actor.getActorID());
+		Event updatedResult = performUpdateEvent(updateEventData, owner, bindingResult);
+		if (bindingResult.hasErrors()) {
+			return RESTResourceResponseData.<Event> badResponse(bindingResult);
+		}
+
+		Event savedEventResult = this.eventRepo.save(updatedResult);
+		Coordinates callerLoc = updateEventData.getCallerCoodinates();
+		Coordinates eventLoc = updateEventData.getEventCoodinates();
+		double distanceFromCaller = GeodeticHelper.getDistanceBetweenCoordinates(callerLoc, eventLoc);
+		System.out.println("DistanceFromCaller: " + distanceFromCaller);
+		
+		return RESTResourceResponseData.createResponse(savedEventResult, HttpStatus.CREATED);
+	}
+
+	private Event performUpdateEvent(UpdateEventData updateEventData, Registrant owner, Errors errors) {
+		Event targetEvent = this.eventRepo.findOne(updateEventData.getEventId());
+		
+		if(!eventContainsOwner(targetEvent,owner,errors)){
+			return targetEvent;
+		}
+		
+		targetEvent.setName(updateEventData.getEventName());
+		targetEvent.setDescription(updateEventData.getEventDescription());
+		targetEvent.setLocation(new Location(updateEventData.getEventCoodinates()));
+		if(!updateEventData.getOccurrencesToAdd().isEmpty()){
+			for(int i=0;i<updateEventData.getOccurrencesToAdd().size();i++){
+				if (!targetEvent.addOccurrence(updateEventData.getOccurrencesToAdd().get(i))){
+					String message = "Cannot update event. Failed to add a coccurrence.";
+					errors.reject("-7", message);
+				}
+			}
+		}
+		if(!updateEventData.getParticipantsToAdd().isEmpty()){
+			for(int i=0;i<updateEventData.getParticipantsToAdd().size();i++){
+				if (!targetEvent.addParticipant(updateEventData.getParticipantsToAdd().get(i))){
+					String message = "Cannot update event. Failed to add a participant.";
+					errors.reject("-7", message);
+				}
+			}
+		}
+		if(!updateEventData.getOwnersToAdd().isEmpty()){
+			for(int i=0;i<updateEventData.getOwnersToAdd().size();i++){
+				if (!targetEvent.addOwner(updateEventData.getOwnersToAdd().get(i))){
+					String message = "Cannot update event. Failed to add an owner.";
+					errors.reject("-7", message);
+				}
+			}
+		}
+		
+		if(!updateEventData.getOccurrencesToRemove().isEmpty()){
+			for(int i=0;i<updateEventData.getOccurrencesToRemove().size();i++){
+				if (!targetEvent.removeOccurrence(updateEventData.getOccurrencesToRemove().get(i))){
+					String message = "Cannot update event. Failed to remove a coccurrence.";
+					errors.reject("-8", message);
+				}
+			}
+		}
+		if(!updateEventData.getParticipantsToRemove().isEmpty()){
+			for(int i=0;i<updateEventData.getParticipantsToRemove().size();i++){
+				if (!targetEvent.removeParticipant(updateEventData.getParticipantsToRemove().get(i))){
+					String message = "Cannot update event. Failed to remove a participant.";
+					errors.reject("-8", message);
+				}
+			}
+		}
+		if(!updateEventData.getOwnersToRemove().isEmpty()){
+			for(int i=0;i<updateEventData.getOwnersToRemove().size();i++){
+				if (!targetEvent.removeOwner(updateEventData.getOwnersToRemove().get(i))){
+					String message = "Cannot update event. Failed to add an owner.";
+					errors.reject("-8", message);
+				}
+			}
+		}
+
+		Category category = this.categoryRepo.findByName(updateEventData.getEventCategory()).get(0);
+		targetEvent.setCategory(category);
+
+		return targetEvent;
+	}
+	
+	private boolean eventContainsOwner(Event targetEvent, Registrant owner, Errors errors) {
+		if(!targetEvent.getOwners().contains(owner)){
+			String message = "Cannot update event. The request Registrant is not the event owner.";
+			errors.reject("-9", message);
+			return false;
+		}
+		return true;
+	}
+
 	@RequestMapping(value = "/rest/events/userJoined")
 	public ResponseEntity<RESTPaginatedResourcesResponseData<Event>> getJoinedEventsList(HttpServletRequest request, BindingResult bindingResult){
 		
@@ -257,10 +385,10 @@ public class EventsController {
 		}
 
 		Actor actor = ActorStateUtility.retrieveActorFromRequest(request);
-		Registrant participant = this.regRepo.findOne(actor.getActorID());
+		Registrant owner = this.regRepo.findOne(actor.getActorID());
 		
 		// Adding user to participant list
-		Event joinedEvent = removeEvent(removeEventData, participant, bindingResult);
+		Event joinedEvent = removeEvent(removeEventData, owner, bindingResult);
 		if (bindingResult.hasErrors()) {
 			return RESTResourceResponseData.<Event> badResponse(bindingResult);
 		}
@@ -268,21 +396,15 @@ public class EventsController {
 		return RESTResourceResponseData.createResponse(joinedEvent, HttpStatus.OK);
 	}
 
-	private Event removeEvent(EventIdData removeEventData, Registrant participant, Errors errors) {
+	private Event removeEvent(EventIdData removeEventData, Registrant owner, Errors errors) {
 		
 		//TODO: add an error check in case event or participant are not found
 		Long eventId = removeEventData.getEventId();
 		Event targetEvent = eventRepo.findOne(eventId);
+		if(!eventContainsOwner(targetEvent,owner,errors)){
+			return targetEvent;
+		}
 		eventRepo.delete(targetEvent);
-//		for (Registrant user : targetEvent.getParticipants()) {
-//		     user.removeJoinedEvent(targetEvent);
-//		}
-//		for (Registrant user : targetEvent.getOnwers()) {
-//		     user.removeOwnedEvent(targetEvent);
-//		}
-//		for (Registrant user : targetEvent.getSubscribers()) {
-//		     user.removeSubscribedEvent(targetEvent);
-//		}
 
 		return targetEvent;
 	}
